+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## [14.4.1] (melonJS 2) - _2023-01-xx_
+## [14.5.0] (melonJS 2) - _2023-01-xx_
+
+### Changed
+- Pool: simplified and optimized the pool `pull` method
 
 ### Fixed
 - Tiled: fix format deprecation warning if map was not created with Tiled and format version is not available
@@ -1407,7 +1410,7 @@
 - first public version with alex4 !
 - www.melonjs.org is live :)
 
-[14.4.1]: https://github.com/melonjs/melonJS/compare/14.4.1...14.4.0
+[14.5.0]: https://github.com/melonjs/melonJS/compare/14.5.0...14.4.0
 [14.4.0]: https://github.com/melonjs/melonJS/compare/14.4.0...14.3.0
 [14.3.0]: https://github.com/melonjs/melonJS/compare/14.3.0...14.2.0
 [14.2.0]: https://github.com/melonjs/melonJS/compare/14.2.0...14.1.2